requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
SQLAlchemy==2.0.36
pyodbc==5.2.0
numpy==2.1.3
scikit-learn==1.5.2
